refactor(mga): extract throw_timing_classifier from classifier_service
classifier_service.py was 2029 LOC (god-module flagged in PR #748).
classify_throw_timing_from_frames + _THROW_TIMING_SCHEMA_DOC are a
self-contained Claude code path (own prompt, own schema, no DB, no slug
resolution) and move cleanly into a sibling module — dropping the parent
file by ~380 LOC to 1649.

Re-exported from classifier_service so existing callers
(throw_localizer, ingestion_orchestrator, micro_clip_generator,
landing_clip_generator, technique_extractor, landing_clip_backfill) do
NOT need import changes. Shared helpers (_GAME_VISUAL_CUES,
_strip_json_fences, _validate_grid_index) stay in classifier_service and
get re-imported by the new file.

test_throw_timing_classifier.py's mock.patch targets updated to the new
module path (settings + anthropic.Anthropic). test_throw_localizer.py
needs no change — it patches at the throw_localizer import site, which
keeps working via the re-export.

Full backend suite (676 tests) green. classifier_service.py still over
1000 LOC; remaining splits (grid_classifier, technique_classifier,
_reference_data) tracked in TECH_DEBT.md.
